requirements.txt
# Core Trading Libraries
kiteconnect==4.3.0
pandas==2.1.4
numpy==1.26.2

# Technical Analysis
ta-lib==0.4.28
pandas-ta==0.3.14b0

# Data Sources
yfinance==0.2.33
nsepy==0.8

# Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0

# Database & Storage
sqlalchemy==2.0.23

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging & Monitoring
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
requests==2.31.0
pytz==2023.3
schedule==1.2.0

# Backtesting
backtrader==1.9.78.123

# Web Framework (for Dashboard - Optional)
flask==3.0.0
